README.md
Objectives:
              The main objective of the java project on employee management system is to manage the details of salary, employee,experience,Informations.It manages all the information about salary, Informations.The project is totally built at administrative end and thus only the administrator is guaranteed the access.The purpose of the project is to build an application program to reduce the manual work  for managing the salary,employee. It tracks all the details about the employees who were working there.

Functionalities provided by this project:
By using this project we can able to create new employee, we can able to Increase or decrease the salary that we have given while adding the employee,we can able to display all Information right from the starting , we can directly decrement the salary of an employee to zero.we can show menu options , we can change status , we can directly exit the program.we can also manage the information of salary.

Report Generation
It generates the report on Salary, type of employee,and about the editing of whole employee detail. It provides filter reports on employee and other informations.This application also provides sub categories of the type of employees.